# [EC] Unify point doubling for P-256/384/521