--- a/crypto/fipsmodule/ec/ec_nistp.h
+++ b/crypto/fipsmodule/ec/ec_nistp.h
@@ -1,0 +1,66 @@
+/*
+------------------------------------------------------------------------------------
+ Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+ SPDX-License-Identifier: Apache-2.0 OR ISC
+------------------------------------------------------------------------------------
+*/
+#ifndef NISTP_H
+#define NISTP_H
+
+#include <openssl/target.h>
+
+#include <stdint.h>
+
+// We have two implementations of the field arithmetic for NIST curves:
+//   - Fiat-crypto
+//   - s2n-bignum
+// Both Fiat-crypto and s2n-bignum implementations are formally verified.
+// Fiat-crypto implementation is fully portable C code, while s2n-bignum
+// implements the operations in assembly for x86_64 and aarch64 platforms.
+// If (1) x86_64 or aarch64, (2) linux or apple, and (3) OPENSSL_NO_ASM is not
+// set, s2n-bignum path is capable.
+#if !defined(OPENSSL_NO_ASM) &&                                                \
+    (defined(OPENSSL_LINUX) || defined(OPENSSL_APPLE)) &&                      \
+    ((defined(OPENSSL_X86_64) && !defined(MY_ASSEMBLER_IS_TOO_OLD_FOR_AVX)) || \
+     defined(OPENSSL_AARCH64))
+#  define EC_NISTP_USE_S2N_BIGNUM
+#  define EC_NISTP_USE_64BIT_LIMB
+#else
+// Fiat-crypto has both 64-bit and 32-bit implementation.
+#  if defined(BORINGSSL_HAS_UINT128)
+#    define EC_NISTP_USE_64BIT_LIMB
+#  endif
+#endif
+
+#if defined(EC_NISTP_USE_64BIT_LIMB)
+typedef uint64_t felem_limb;
+#else
+typedef uint32_t felem_limb;
+#endif
+
+// This is the struct that holds pointers to implementations of field
+// arithmetic functions for specific curves. It is meant to be used
+// in higher level functions like this:
+//   void point_double(nistp_felem_methods *ctx, ...) {
+//     ctx->add(...);
+//     ctx->mul(...);
+//   }
+// This makes the functions reusable between different curves by simply
+// providing an appropriate methods object.
+typedef struct {
+  void (*add)(felem_limb *c, const felem_limb *a, const felem_limb *b);
+  void (*sub)(felem_limb *c, const felem_limb *a, const felem_limb *b);
+  void (*mul)(felem_limb *c, const felem_limb *a, const felem_limb *b);
+  void (*sqr)(felem_limb *c, const felem_limb *a);
+} nistp_felem_methods;
+
+
+void nistp_point_double(nistp_felem_methods *ctx,
+                        felem_limb *x_out,
+                        felem_limb *y_out,
+                        felem_limb *z_out,
+                        const felem_limb *x_in,
+                        const felem_limb *y_in,
+                        const felem_limb *z_in);
+#endif // NISTP_H
+