--- a/crypto/fipsmodule/ec/p256.c
+++ b/crypto/fipsmodule/ec/p256.c
@@ -167,11 +167,12 @@ static void fiat_p256_inv_square(fiat_p256_felem out,
   fiat_p256_square(out, ret);  // 2^256 - 2^224 + 2^192 + 2^96 - 2^2
 }
 
-static ec_nistp_felem_meth p256_felem_methods = {
-                                fiat_p256_add,
-                                fiat_p256_sub,
-                                fiat_p256_mul,
-                                fiat_p256_square };
+DEFINE_LOCAL_DATA(ec_nistp_felem_meth, p256_felem_methods) {
+    out->add = fiat_p256_add;
+    out->sub = fiat_p256_sub;
+    out->mul = fiat_p256_mul;
+    out->sqr = fiat_p256_square;
+}
 
 static void fiat_p256_point_double(fiat_p256_felem x_out,
                                    fiat_p256_felem y_out,
@@ -179,7 +180,7 @@ static void fiat_p256_point_double(fiat_p256_felem x_out,
                                    const fiat_p256_felem x_in,
                                    const fiat_p256_felem y_in,
                                    const fiat_p256_felem z_in) {
-  ec_nistp_point_double(&p256_felem_methods, x_out, y_out, z_out, x_in, y_in, z_in);
+  ec_nistp_point_double(p256_felem_methods(), x_out, y_out, z_out, x_in, y_in, z_in);
 }
 
 // fiat_p256_point_add calculates (x1, y1, z1) + (x2, y2, z2)